backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Utilities
python-dateutil>=2.8.2